Use gamepad to control the power for elevator

+++ src/Commands/WinchUpDown.cpp
@@ -23,7 +23,7 @@ void WinchUpDown::Initialize()
 // Called repeatedly when this Command is scheduled to run
 void WinchUpDown::Execute()
 {
-	int temp_power = m_power;
+	double temp_power = m_power;
 
 	if(m_isUp) {
 		Robot::winch->MoveUp(temp_power);
